adding most of the position files

+++ src/main/java/org/llap/prosper/trader/repositories/PositionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 
 	@Repository
 	public interface PositionRepository extends
-	        JpaRepository<Position, Long> {
+	        JpaRepository<Position, Integer> {
 	    @Query("SELECT p FROM Position p ")
 	    List<Position> findAll();
 	}
